backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
cohere>=4.4.3
qdrant-client>=1.7.0
python-dotenv>=1.0.0
pytest>=7.4.0
fastapi>=0.104.1
uvicorn>=0.24.0
slowapi>=0.1.9